Adds a failing test for #3060
